Convert agreement to float.

+++ scripts/evaluate_story_generation_tasks.py
@@ -138,7 +138,7 @@ def worker_agreement(output_dir, questions, story_df):
                 for index, row in model_df.iterrows():
                     # coder, item, labels
                     #print(row)
-                    all_triples.append((row[WORKER_COL], f"{row['model_type']}_{row[HIT_COL]}", int(row[f"{q}_ranking"])))
+                    all_triples.append((row[WORKER_COL], f"{row['model_type']}_{row[HIT_COL]}", float(row[f"{q}_ranking"])))
 
                 res_dict = annotation_agreement(all_triples, f"ranking")
                 res_dict["worker_1"] = "all"
@@ -158,12 +158,12 @@ def worker_agreement(output_dir, questions, story_df):
                     for index, row in worker_one_df.iterrows():
                         # coder, item, labels
                         #print(row)
-                        triples.append((w[0], f"{row['model_type']}_{row[HIT_COL]}", int(row[f"{q}_ranking"])))
+                        triples.append((w[0], f"{row['model_type']}_{row[HIT_COL]}", float(row[f"{q}_ranking"])))
 
                     for index, row in worker_two_df.iterrows():
                         #print(row)
                         # coder, item, labels
-                        triples.append((w[1], f"{row['model_type']}_{row[HIT_COL]}", int(row[f"{q}_ranking"])))
+                        triples.append((w[1], f"{row['model_type']}_{row[HIT_COL]}", float(row[f"{q}_ranking"])))
 
                     res_dict = annotation_agreement(triples, f"ranking")
                     res_dict["worker_1"] = w[0]
